requirements.txt
# Core dependencies for inference
docaligner-docsaid>=1.1.0
opencv-python>=4.8.0
numpy>=1.24.0,<2.0

# Training dependencies (optional - only needed for custom training)
torch>=2.0.0
torchvision>=0.15.0
tqdm>=4.65.0
tensorboard>=2.13.0

# Development dependencies (optional)
pytest>=7.4.0
black>=23.7.0
flake8>=6.1.0
